Refactor: adjusting location of POST validations

+++ src/main/java/com/electricitybill/service/impl/PosteServiceImpl.java
@@ -3,10 +3,12 @@ package com.electricitybill.service.impl;
 import com.electricitybill.entity.PosteEntity;
 import com.electricitybill.repository.PosteRepository;
 import com.electricitybill.service.PosteService;
+import jakarta.persistence.NoResultException;
 import lombok.AllArgsConstructor;
 
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Objects;
 
 @AllArgsConstructor
 public class PosteServiceImpl implements PosteService {
@@ -15,11 +17,17 @@ public class PosteServiceImpl implements PosteService {
 
     @Override
     public List<PosteEntity> findAll() {
+        if (posteRepository.findAll().isEmpty()) {
+            throw new NoResultException("Search not found!");
+        }
         return posteRepository.findAll();
     }
 
     @Override
     public PosteEntity findById(int id) {
+        if (Objects.isNull(posteRepository.findById(id))) {
+            throw new NoResultException("Individual search not found!");
+        }
         return posteRepository.findById(id);
     }
 
@@ -30,11 +38,17 @@ public class PosteServiceImpl implements PosteService {
 
     @Override
     public boolean update(int id, PosteEntity entity) throws SQLException {
+        if (Objects.isNull(posteRepository.findById(id))) {
+            throw new NoResultException("Post not found!");
+        }
         return posteRepository.update(id, entity);
     }
 
     @Override
     public boolean delete(int id) throws SQLException {
+        if (Objects.isNull(posteRepository.findById(id))) {
+            throw new NoResultException("Post not found!");
+        }
         return posteRepository.delete(id);
     }
 }
